loggerdb revamp:
We now support resuming on an old database, so long as it falls into the
previous genconfig.CandleSize period.

A new issue and TODO:
the avarice threading is still constant. What that means, is that if you
have a genconfig.CandleSize of 10, kill avarice, start it 5 minutes
later, then you will have a 15 minute candle on startup.

Also, even though we're no longer getting rid of this table, indicators
are still populating lists (not rows). This also means that RS gain avgs
are only read from the last and second to last price_list entries from
an avarice resume.