Revert to old Trade Indicators format:
ConfigObj can keep a nested list format immediately after writing,
but not after re-opening the file. Validate checks malform the nested
list, so treating them as strings and using ast seems like a decent solution.

This will be simplified in webconfigure.py with checkboxes, and then converting
the list to a string.